requirements.txt
alembic==1.12.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cattrs==23.1.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.2
executing==1.2.0
fastjsonschema==2.18.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
fonttools==4.42.1
fqdn==1.5.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.4
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
jupyterthemes==0.20.0
kiwisolver==1.4.5
lesscpy==0.15.1
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==3.0.1
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.2
notebook_shim==0.2.3
numpy==1.26.0
overrides==7.4.0
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
ply==3.11
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2-binary==2.9.7
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.30.2
requests==2.31.0
requests-cache==1.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.21
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
uri-template==1.3.0
url-normalize==1.4.3
urllib3==2.0.4
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.7
WTForms==3.0.1
